2-do_deploy_web_static.py_22:02

+++ 2-do_deploy_web_static.py
@@ -15,17 +15,20 @@ def do_deploy(archive_path='none'):
     """
     Function to pack files in web_static into a compressed file
     """
+    if (run("test -e {}".format(archive_path)).failed):
+        return False
+
     try:
         with cd("/tmp"):
             put("{}".format(archive_path), "/tmp/")
 
-            output_object = run("echo $(ls -t /tmp | awk 'NR==1')")
+            output_object = run("ls -t /tmp | awk 'NR==1'")
             archive_name = output_object.stdout
             raw_name = run("echo \"{}\" | cut -d '.' -f 1".format(archive_name)).stdout
 
             run("mkdir -p /data/web_static/releases")
-            run("tar -xvf /tmp/{} -C /data/web_static/releases"\
-                .format(archive_name))
+            run("tar -xvf /tmp/{} -C /data/web_static/releases/{}"\
+                .format(archive_name, raw_name))
             run("rm -f {}".format(archive_name))
             run("rm -f /data/web_static/current")
             run("ln -s -f /data/web_static/releases/{} /data/web_static/current".format(raw_name))
